refactor: modify landing page query
As it was originally, the `.findOneAndUpdate()` query would set the `saved` field to false on all updated articles. This caused saved articles not to persist after a scrape operation.

To address this issue, I removed the `saved` field from the
`.findOneAndUpdate()` query; however, this caused all new scraped articles to persist without the `saved` field because this is an update, rather than a create operation.

The simplest way to addres this was to modify the landing page query such that it queried on the `saved` field being set to false or the `saved` field being non-existent.

In a future update, I may look at modifying the original query that persists articles in the database. I may want to include a pre-hook that detemines if the document already exists before performing the update. (See https://masteringjs.io/tutorials/mongoose/findoneandupdate and https://stackoverflow.com/questions/16882938/how-to-check-if-that-data-already-exist-in-the-database-during-update-mongoose/16883572.) I could also turn the query into a `.create()` operation.

For now, this solution encapsulated in this commit appears to work.