fix: modify clear articles (`deleteMany()`) query
Due to a recent change in how news articles are persisted in the database, it appeared the Clear Articles functionality was not working. However, upon testing, the `deleteMany()` query was functioning correctly; it simply checked if the `saved` field was `false`. If so,
the news article was deleted. The query now checks if `saved` is `false` OR if the `saved` field is nonexistent. In either case, the news article is deleted.